+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Updated
+
+- `def` syntax annotation updated.
+
 ## [0.0.7] - 2023-05-15
 
 ### Added